+++ includes/vectors.h
@@ -3,15 +3,15 @@
 # include "math.h"
 
 typedef struct		t_vector {
-		double		v_x;
-		double		v_y;
-		double		v_z;
+		float		v_x;
+		float		v_y;
+		float		v_z;
 }					s_vector;
 
 typedef struct		t_point {
-		double		p_x;
-		double		p_y;
-		double		p_z;
+		float		p_x;
+		float		p_y;
+		float		p_z;
 }					s_point;
 
 typedef struct		t_ray {
@@ -20,7 +20,7 @@ typedef struct		t_ray {
 }					s_ray;
 
 typedef struct		t_cam_to_w {
-		double		matrix[4][4];
+		float		matrix[4][4];
 }					s_cam_to_w;
 
 /**
@@ -32,15 +32,15 @@ typedef struct		t_cam_to_w {
  ** new_vector = create new vector
 **/
 
-s_point			*new_point(double x, double y, double z);
-double			vector_scalar_mult(s_vector *a, s_vector *b);
-double			vector_length(s_vector *v);
-s_vector		*vector_normalise(s_vector *v, double len);
+s_point			*new_point(float x, float y, float z);
+float			vector_scalar_mult(s_vector *a, s_vector *b);
+float			vector_length(s_vector *v);
+s_vector		*vector_normalise(s_vector *v, float len);
 s_vector		*cross_prod(s_vector *a, s_vector *b);
 s_vector		*add_vectors(s_vector *a, s_vector *b);
 s_vector		*subs_vectors(s_vector *a, s_vector *b);
-s_vector		*vector_by_scalar(s_vector *a, double num);
-s_vector		*new_vector(double x, double y, double z);
+s_vector		*vector_by_scalar(s_vector *a, float num);
+s_vector		*new_vector(float x, float y, float z);
 s_vector		*vector_from_points(s_point *p1, s_point *p2);
 
 s_cam_to_w		matrix_place(s_vector *coor, s_vector *dir);
