+++ parser/utils.c
@@ -1,10 +1,10 @@
 #include "../includes/MiniRT.h"
 
-double		d_atoi(char *str)
+float		d_atoi(char *str)
 {
 	int 	i;
 	int 	sign;
-	double	res;
+	float	res;
 
 	i = 0;
 	sign = 1;
@@ -19,19 +19,19 @@ double		d_atoi(char *str)
 	while (str[i] >= '0' && str[i] <= '9' && str[i])
 		++i;
 	if (!str[i])
-		return (res * (double)sign);
+		return (res * (float)sign);
 	if (str[i] == '.')
 	{
 		++i;
 		res += parse_d_part(&str[i]);
 	}
-	return (res * (double)sign);
+	return (res * (float)sign);
 }
 
-double 		parse_int_part(char *str)
+float 		parse_int_part(char *str)
 {
 	int		i;
-	double	res;
+	float	res;
 
 	i = 0;
 	res = 0;
@@ -43,10 +43,10 @@ double 		parse_int_part(char *str)
 	return (res);
 }
 
-double 		parse_d_part(char *str)
+float 		parse_d_part(char *str)
 {
 	int		i;
-	double	res;
+	float	res;
 
 	i = 0;
 	res = 0;
