+++ vectors_funcs/vector_length.c
@@ -1,6 +1,6 @@
 #include "../includes/MiniRT.h"
 
-double	vector_length(s_vector *v)
+float	vector_length(s_vector *v)
 {
 	return (sqrt(v->v_x * v->v_x + v->v_y * v->v_y + v->v_z * v->v_z));
 }
